+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpForEachStatement.cs
@@ -5,5 +5,6 @@ public class CSharpForEachStatement : CSharpStatementBlock
     public CSharpForEachStatement(string iterationVariable, string sourceCollection)
     {
         Text = $"foreach (var {iterationVariable} in {sourceCollection})";
+        BeforeSeparator = CSharpCodeSeparatorType.EmptyLines;
     }
 }